feature: 插件访问数据量相关字段及统计字段改为float类型，降低内存使用。 feature: 增加对mysql数据库的支持，默认数据库切换为mysql。 bug fix: getResponseTimeAvg方法使用了无法跨数据库的数据类型，导致在mysql下无法使用。

+++ source/cordwood-console/src/main/java/ren/hankai/cordwood/console/persist/PluginRequestRepositoryCustom.java
@@ -40,7 +40,7 @@ public interface PluginRequestRepositoryCustom {
    * @author hankai
    * @since Dec 12, 2016 1:40:34 PM
    */
-  Double getUserPluginTimeUsageAvg(String userEmail, Date beginTime, Date endTime);
+  Float getUserPluginTimeUsageAvg(String userEmail, Date beginTime, Date endTime);
 
   /**
    * 所有插件的流量和（请求+响应数据字节数）。
@@ -51,7 +51,7 @@ public interface PluginRequestRepositoryCustom {
    * @author hankai
    * @since Dec 12, 2016 1:41:53 PM
    */
-  Long getPluginTotalDataBytes(Date beginTime, Date endTime);
+  Float getPluginTotalDataBytes(Date beginTime, Date endTime);
 
   /**
    * 获取用户插件流量和。
@@ -63,7 +63,7 @@ public interface PluginRequestRepositoryCustom {
    * @author hankai
    * @since Dec 12, 2016 1:42:48 PM
    */
-  Long getUserPluginDataBytes(String userEmail, Date beginTime, Date endTime);
+  Float getUserPluginDataBytes(String userEmail, Date beginTime, Date endTime);
 
   /**
    * 获取按插件分组汇总后的访问统计信息。
